Liberate topology generators from Geometry
